+++ rest_framework_services/types/__init__.py
@@ -5,11 +5,14 @@ These are intentionally framework-agnostic data carriers. They live outside
 inputs/outputs are not coupled to the mutation helpers themselves.
 """
 
+from rest_framework_services.types.agent_field import AGENT, AgentField
+from rest_framework_services.types.agent_projection import AgentProjection
 from rest_framework_services.types.argument_binding import ArgumentBinding
 from rest_framework_services.types.change_result import ChangeResult
 from rest_framework_services.types.child_collection_change import ChildCollectionChange
 from rest_framework_services.types.child_spec import ChildSpec
 from rest_framework_services.types.dispatch_result import DispatchResult
+from rest_framework_services.types.field_audience import FieldAudience
 from rest_framework_services.types.field_change import FieldChange
 from rest_framework_services.types.forward_relation_spec import ForwardRelationSpec
 from rest_framework_services.types.generic_relation_spec import GenericRelationSpec
@@ -50,6 +53,9 @@ from rest_framework_services.types.validate_channel_names import validate_channe
 from rest_framework_services.types.view_hooks import ViewHooks
 
 __all__ = [
+    "AGENT",
+    "AgentField",
+    "AgentProjection",
     "DEFAULT_JSON_SCHEMA_REGISTRY",
     "RESERVED_POOL_SEEDS",
     "UNSET",
@@ -83,6 +89,7 @@ __all__ = [
     "RelationPhase",
     "RelationSpec",
     "ReverseOneToOneSpec",
+    "FieldAudience",
     "SelectorKind",
     "SelectorSpec",
     "ServiceSpec",